+++ .gitignore
@@ -231,3 +231,5 @@ fabric.properties
 
 # Android studio 3.1+ serialized cache file
 .idea/caches/build_file_checksums.ser
+
+runtime.txt
